pyproject.toml
[project]
name = "rag-deep-dive"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "colorama>=0.4.6",
    "deepeval==3.1.0",
    "faiss-cpu==1.11.0",
    "langchain-community==0.3.25",
    "langchain-google-genai>=2.1.7",
    "langchain-groq>=0.3.6",
    "langchain-huggingface>=0.3.0",
    "nltk>=3.9.1",
    "pandas>=2.3.1",
    "pymupdf==1.26.1",
    "pypdf==5.6.0",
    "python-dotenv==1.1.0",
    "rank-bm25==0.2.2",
    "sentence-transformers>=5.0.0",
]
